A_Barrier object class
Contains float variable "HP", and 5 scene objects parented in series Breakable1 through Breakable5.

As HP is reduced down to 0, it will "activate" different Breakable levels.

When activated, all child objects will do one of two things, depending on the type:
If it is a collider, it will be deleted.
If it is a cube or other primitive, it will have physics enabled but will only collide with the ground and other subobjects.
Depending on the source of damage, an impulse will be applied to each primitive.